API documentation

Data Sources

Data from arbitrary file formats can be used as long as they return an array of values when provided with a string.

CrystalInfoContainers.DataSource — Type

A DataSource, when provided with a string referring to a data name from a namespace known to the data source, returns an array of values that are either Native (already Julia types) or Text (requiring interpretation to create Julia types).

CrystalInfoContainers.TypedDataSource — Type

TypedDataSource(data,dictionary)

A TypedDataSource is a DataSource that returns items with the correct type and aliases resolved, as specified in the associated CIF dictionary. The dictionary also provides a namespace.